refactor: remove unused last-trip tracking
At the time this was implemented almost a year ago, it was thought that
we would begin using it soon as part of the "real-time departure states"
project. This project has both evolved and been deprioritized since. In
keeping with our general stance of not keeping things in the codebase
which are not actively being used (especially something like this which
consumes non-trivial resources in production), this can be removed for
now. The eventual future implementation of last-trip tracking may look
slightly different, though we can also always refer back to this code in
the history of the repo.