chore: disable unneeded telemetry logging
This method of sending telemetry was regularly overloading the Splunk
log handler and causing more important logs to be dropped (in particular
DeviceMonitor, which we rely on for alerting us to screen outages). We
have not actually used these log spans to investigate anything in a long
time, so we'll remove them until we may need them again.