fix: missing "due to" in Pre-Fare alert audio
c945169 fixed an issue where audio readouts for Pre-Fare takeover alerts
did not include the cause of an alert (e.g. "maintenance") as intended.
This fix included a bug of its own where the phrase "due to" was not
included as it should have been (e.g. "due to maintenance"). This fixes
that bug.