fix: interpreting alerts as directional
`Alert.direction_id/1` would incorrectly return a non-nil value if
there happened to be any stopping locations (which would always have
a non-nil direction) ahead of the first parent station in the informed
entities. Since entities are never explicitly sorted, this could
sometimes be the case and sometimes not, even for the same alert.

The immediate fix is to only consider parent stations, though this
approach is still flawed, as is the entire notion that one can assign a
single "direction" to an alert. It should work well enough for how
alert data is currently structured and the kinds of alerts we are
likely to see, though.