fix: rework text auto-sizing hook
`useTextResizer` had a known bug where the "sizing step" could stop one
step short of the correct one, leaving the element visibly overflowing.
This was most likely due to its use of `useEffect`, which runs async
with DOM updates and is not guaranteed to "see" the DOM in a particular
state. In React 17 it happened that this bug only occurred rarely, but
in React 18, presumably due to internal changes or optimizations in the
render cycle, it now occurs more often.

The main change necessary here is to switch to `useLayoutEffect`,
which *is* guaranteed to see the DOM after it has been updated but
before the browser repaints. However, while we're here, we can make
some further improvements:

* Check for overflow directly rather than requiring the hook user to
  specify a maximum pixel height. This allows size constraints to be
  expressed in CSS where they belong.

* Remove the `isDone` return value. No users of the hook actually needed
  this other than to know when the final resizing step had been reached,
  which can also be checked more directly.