fix: header sizing and alignment
Using the correct fonts for headers exposed an issue with the recent
auto-sizing refactor where `lcd_common` screens were always in
the "small" state even when there was room for a "large" header.

The problem was a combination of the text using `line-height: 1`, which
cuts off some of the space intended for descenders, plus the container
not having any specified size and conforming to the size of the text.
When this is the case, the container "has overflow" because if it had
`overflow: hidden` the shortened line-height would cause the descenders
to be cut off.

The solution is to allow this container to be 100% of the height of the
header slot, which has no downsides and also allows a simplification of
how header elements (icon and text) are vertically aligned.