refactor: fix Rules of Hooks violations
We added linting for this some time ago, but the pattern for files it
would check was mistakenly set to `**/*.{ts, tsx}`, which only matches
`.ts` files (and `. tsx` files, space included... if we had any).

With this pattern updated to correctly match `.tsx` files, several
errors are uncovered. This fixes or ignores them as best we can.