fix: crash in departure building on nil times
Sometimes Commuter Rail has "scheduled" stop time updates with null
arrival and departure time (though technically this isn't valid GTFS-RT
data...). When this happens, departure building crashes when it tries
to pass `nil` to `DateTime.compare/2` as part of sorting the departures
by time.

Handle this in a way that will also accommodate displaying
skipped/cancelled departures, since we know we'll want to do that in
the future; departure building continues to filter them out for now.