chore: don't run all screen variants by default
This was a feature intended for the DUP departures variant that never
saw "real" use (with a non-trivial candidate generator) until now, with
the introduction of the Pre-Fare `new_elevators` variant. The original
idea was a sort of automated smoke testing — if there were any crashes
in the new logic caused by unforeseen screen configurations or
real-world conditions, we'd see them in production, but before the new
logic was actually hooked up to real screens. However, a side effect of
this is that every screen variant multiplies the "effective" number of
screens of its type, putting additional load on the app. This is likely
what is causing our recent high-memory alerts.

For now, just remove this option. We might consider adding it back in
the future on a random-sampling basis (e.g. on only 1% of all requests).