feat: MBTA Go QR codes for new elevator status
In the case where one or more elevators are closed at the screen's
station, the QR code in the widget should point to the appropriate
`go.mbta.com` URL for either the alert (if only one) or the stop in
general. This should redirect riders to either the relevant view in
MBTA Go if they have it installed, or the equivalent page on mbta.com
if they don't.