feat: promote new elevator status widget to live
* Delete all files associated with the existing Elevator Status widget.
* Remove the variant candidate generator and use the new widget in the
  default candidate generator.
* Remove "new" naming from module names, CSS classes, etc. now that the
  non-"new" name is freed up.

N.B. this commit does not rename widget-related *files* with "new" in
their names, because the combination of deleting an old file and moving
a new file into its place is not properly detected as a move by Git,
resulting in scrambled history. The commit immediately following this
one will perform this move.