feat: paging indicator for Pre-Fare flex zone
This flex zone was never given a paging indicator, presumably because
the elevator status widget also had one, and two independent paging
indicators on the same screen would have looked strange/confusing. Now
that the new non-paged elevator status widget is implemented, the flex
zone is the only paged slot on the right screen, so it is free to adopt
the paging indicator.