fix(ElevatorStatus): including "nearby" closures
The desired behavior for this widget is that closures with nearby
redundancy are ignored when selecting a "state", but displayed (whatever
that means for the given state) once we are in it.

The "closed elsewhere" state was correctly following this rule, but
the "closed here" state was incorrectly filtering out closures with
nearby redundancy from being displayed at all.